requirements.txt
streamlit>=1.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
bcrypt>=4.1.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0

